% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.coerce_progress}
\alias{.coerce_progress}
\title{Coerce a progress argument into a plyr-compatible value}
\usage{
.coerce_progress(x = NULL)
}
\description{
Coerce a progress argument into a plyr-compatible value
}
\keyword{internal}
