% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{ggneuron}
\alias{ggneuron}
\title{Create a ggplot2 Visualisation of Neuron Objects}
\usage{
ggneuron(
  x,
  volume = NULL,
  info = NULL,
  rotation_matrix = NULL,
  cols1 = c("turquoise", "navy"),
  cols2 = c("grey75", "grey50"),
  alpha = 0.5,
  title.col = "darkgrey",
  ...
)
}
\arguments{
\item{x}{A 'neuron', 'neuronlist', 'mesh3d', or 'hxsurf' object to be visualised.}

\item{volume}{a brain/neuropil volume to be plotted in grey, for context.
Defaults to NULL, no volume plotted.}

\item{info}{Optional. A string to be used as the plot title.}

\item{rotation_matrix}{An optional 4x4 rotation matrix to apply to the neuron coordinates.}

\item{cols1}{Colour for the lowest Z values. Default is "turquoise".}

\item{cols2}{Colour for the highest Z values. Default is "navy".}

\item{alpha}{Transparency of the neuron visualisation. Default is 0.5.}

\item{title.col}{Colour of the plot title. Default is "darkgrey".}

\item{...}{Additional arguments passed to geom_neuron().}
}
\value{
A ggplot2 object representing the neuron visualisation.
}
\description{
This function creates a complete ggplot2 visualisation for neuron objects,
including 'neuron', 'neuronlist', 'mesh3d', and 'hxsurf' objects. It sets up
a minimal theme and applies consistent styling to the plot.
}
\details{
This function wraps around geom_neuron() to create a complete plot with a
consistent, minimal theme. It removes axes, legends, and other extraneous
elements to focus on the neuron visualisation itself.
}
\examples{
\dontrun{
library(nat.ggplot)

# Visualise neurons with brain volume as context
ggneuron(banc.skels,
         volume = banc.brain_neuropil,
         rotation_matrix = banc_view)

# Visualise the brain neuropil alone
ggneuron(banc.brain_neuropil,
         rotation_matrix = banc_view,
         cols1 = c("lightblue", "darkblue"))

# Visualise split neurons with custom colours
ggneuron(bc.neurons.flow,
         volume = banc.brain_neuropil,
         rotation_matrix = banc_view,
         info = "LHPD2a1 neurons with axon/dendrite split")

# Visualise neuron meshes
ggneuron(banc.meshes[[1]],
         rotation_matrix = banc_view,
         cols1 = c("purple", "magenta"),
         alpha = 0.8)
}

}
\references{
Jefferis, G. S. X. E., Potter, C. J., Chan, A. M., Marin, E. C., Rohlfing, T.,
Maurer, C. R., & Luo, L. (2007). Comprehensive maps of Drosophila higher
olfactory centers: Spatially segregated fruit and pheromone representation.
\emph{Cell}, 128(6), 1187-1203. \doi{10.1016/j.cell.2007.01.040}

Schneider-Mizell, C. M., Gerhard, S., Longair, M., Kazimiers, T., Li, F.,
Zwart, M. F., Champion, A., Midgley, F. M., Fetter, R. D., Saalfeld, S.,
& Cardona, A. (2016). Quantitative neuroanatomy for connectomics in Drosophila.
\emph{eLife}, 5, e12059. \doi{10.7554/eLife.12059}
}
\seealso{
\code{\link{geom_neuron}} for the underlying geom used in this function.
}
