% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rgl_view}
\alias{rgl_view}
\title{Get Current RGL View Parameters for Use as Rotation Matrix}
\usage{
rgl_view()
}
\value{
A list containing:
\describe{
\item{userMatrix}{A 4x4 rotation matrix that can be used with the
\code{rotation_matrix} parameter in nat.ggplot functions}
\item{zoom}{The current zoom level}
\item{windowRect}{The window dimensions}
}
}
\description{
Captures the current viewing parameters from an active rgl window and returns
them in a format suitable for use as a rotation matrix with nat.ggplot functions.
This allows users to interactively find their desired viewing angle using
\code{plot3d()} and then apply the same view to ggplot2 visualisations.
}
\details{
The typical workflow is:
\enumerate{
\item Plot neurons or brain meshes using \code{plot3d()} from the nat package
\item Interactively rotate the view using the mouse to find the desired angle
\item Call \code{rgl_view()} to capture the current view parameters
\item Use the \code{$userMatrix} component as the \code{rotation_matrix} argument
in nat.ggplot functions like \code{geom_neuron()} or \code{ggneuron()}
}
}
\examples{
\dontrun{
library(nat)
library(nat.ggplot)
library(ggplot2)

# First, plot neurons in 3D and rotate to desired view
plot3d(banc.brain_neuropil, alpha = 0.3)
plot3d(banc.skels)

# Capture the view after rotating with mouse
my_view <- rgl_view()

# Use the captured view in ggplot2
ggplot() +
  geom_neuron(banc.brain_neuropil,
              rotation_matrix = my_view$userMatrix,
              cols = c("grey75", "grey50"),
              alpha = 0.3) +
  geom_neuron(banc.skels,
              rotation_matrix = my_view$userMatrix)
}

}
\seealso{
\code{\link{geom_neuron}}, \code{\link{ggneuron}}
}
